# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DiscipleOS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface DeleteById {
  id: string;
}

export interface Event {
  id: string;
  title: string;
  type: string;
  date: string;
  time: string;
  notes?: string;
  remind?: boolean;
  reminderMinutes?: number;
  repeat?: string;
  repeatWeekdays?: number[];
  repeatUntil?: string;
  timeZone?: string;
  countsTowardRhythm?: boolean;
}

export interface EventInput {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  replacesEventId?: string | null;
  title: string;
  type: string;
  date: string;
  time: string;
  notes?: string;
  remind?: boolean;
  reminderMinutes?: number;
  repeat?: string;
  repeatWeekdays?: number[];
  /** @nullable */
  repeatUntil?: string | null;
  timeZone?: string;
  countsTowardRhythm?: boolean;
}

export interface ReadingPlanInput {
  id: string;
  templateKey?: string;
  [key: string]: unknown;
 }

export interface ReadingCompleteInput {
  planId: string;
  key: string;
  completed: boolean;
}

export interface EventCompleteInput {
  eventId: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  occurrenceDate: string;
  completed: boolean;
}

export interface EventCompleteResponse {
  success: boolean;
  eventId: string;
  occurrenceDate: string;
  completed: boolean;
}

export type MountainRhythmResponseAccessTier = typeof MountainRhythmResponseAccessTier[keyof typeof MountainRhythmResponseAccessTier];


export const MountainRhythmResponseAccessTier = {
  basic: 'basic',
  full: 'full',
} as const;

export type MountainRhythmResponseCategories = {[key: string]: {
  completed: number;
  planned: number;
}};

export type MountainRhythmResponseCommitmentsItem = { [key: string]: unknown };

export interface MountainRhythmResponse {
  success: boolean;
  accessTier: MountainRhythmResponseAccessTier;
  level: string;
  percentage: number;
  completed: number;
  planned: number;
  windowStart: string;
  windowEnd: string;
  message?: string;
  copy: string;
  categories: MountainRhythmResponseCategories;
  commitments: MountainRhythmResponseCommitmentsItem[];
}

export type PushSubscriptionInputKeys = {
  p256dh: string;
  auth: string;
};

export interface PushSubscriptionInput {
  endpoint: string;
  keys: PushSubscriptionInputKeys;
}

export interface TrackEventInput {
  event?: string;
  /** @nullable */
  userId?: string | null;
  [key: string]: unknown;
 }

export type SettingsInputPaceMode = typeof SettingsInputPaceMode[keyof typeof SettingsInputPaceMode];


export const SettingsInputPaceMode = {
  relaxed: 'relaxed',
  standard: 'standard',
  fast: 'fast',
  custom: 'custom',
} as const;

export type SettingsInputReadingOrder = typeof SettingsInputReadingOrder[keyof typeof SettingsInputReadingOrder];


export const SettingsInputReadingOrder = {
  consecutive: 'consecutive',
  randomized: 'randomized',
} as const;

export interface SettingsInput {
  paceMode: SettingsInputPaceMode;
  /**
     * @minimum 1
     * @maximum 60
     */
  customMinutesPerChapter: number;
  /**
     * @minimum 5
     * @maximum 240
     */
  dailyReadingBudget: number;
  /** @pattern ^\d{2}:\d{2}$ */
  preferredReadingTime: string;
  readingOrder: SettingsInputReadingOrder;
}

export interface SettingsResponse {
  success: boolean;
  settings: SettingsInput;
}

export type ListEvents200 = {
  success: boolean;
  events: Event[];
};

export type UpsertEvent200 = {
  success: boolean;
  event: Event;
};

export type ListReadingPlans200 = {
  success: boolean;
  plans: unknown[];
};

export type GetMountainRhythmScoreParams = {
timeZone?: string;
};