# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DiscipleOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                required: [success, events]
    post:
      operationId: upsertEvent
      tags: [events]
      summary: Create or update an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  event:
                    $ref: "#/components/schemas/Event"
                required: [success, event]
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteById"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /events/complete:
    post:
      operationId: markEventComplete
      tags: [events]
      summary: Mark one event occurrence complete or incomplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCompleteInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCompleteResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    DeleteById:
      type: object
      properties:
        id:
          type: string
      required: [id]
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        date:
          type: string
        time:
          type: string
        notes:
          type: string
        remind:
          type: boolean
        reminderMinutes:
          type: integer
        repeat:
          type: string
        repeatWeekdays:
          type: array
          items:
            type: integer
        repeatUntil:
          type: string
        timeZone:
          type: string
        countsTowardRhythm:
          type: boolean
      required: [id, title, type, date, time]
    EventInput:
      type: object
      properties:
        id:
          type: ["string", "null"]
        replacesEventId:
          type: ["string", "null"]
        title:
          type: string
        type:
          type: string
        date:
          type: string
        time:
          type: string
        notes:
          type: string
        remind:
          type: boolean
        reminderMinutes:
          type: integer
        repeat:
          type: string
        repeatWeekdays:
          type: array
          items:
            type: integer
        repeatUntil:
          type: ["string", "null"]
        timeZone:
          type: string
        countsTowardRhythm:
          type: boolean
      required: [title, type, date, time]
...
    EventCompleteInput:
      type: object
      properties:
        eventId:
          type: string
        occurrenceDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        completed:
          type: boolean
      required: [eventId, occurrenceDate, completed]
    EventCompleteResponse:
      type: object
      properties:
        success:
          type: boolean
        eventId:
          type: string
        occurrenceDate:
          type: string
        completed:
          type: boolean
      required: [success, eventId, occurrenceDate, completed]
...